Extract deploy/env.go
